+++ src/main/java/disIND/valueBased/model/SharedModel.java
@@ -374,7 +374,7 @@ public final class SharedModel {
         record LhsLiveDelta( int colId, int round , RoaringBitmap newValues) implements CMCommand {}
         record RhsLiveDelta( int colId, int round , RoaringBitmap newValues) implements CMCommand {}
         record MembershipResult(UnaryPair pair, int round, RoaringBitmap missingValues) implements CMCommand {}
-        record ValueOwnerCandidateStatusUpdate(long epoch, long voSequence, int round, int bucketId,
+        record ValueOwnerCandidateStatusUpdate(long epoch, int voSequence, int round, int bucketId,
                 List<CandidateLocalStatus> statuses) implements CMCommand {
             public ValueOwnerCandidateStatusUpdate {
                 if (voSequence <= 0)
